Add some configs to openai

+++ super-heroes/rest-villains/src/main/java/io/quarkus/workshop/superheroes/villain/VillainService.java
@@ -56,7 +56,8 @@ public class VillainService {
 
     public void deleteVillain(Long id) {
         Villain villain = Villain.findById(id);
-        villain.delete();
+        if (villain != null)
+            villain.delete();
     }
 
 }
